Swagger

See merge request lamenu/backend!16

+++ src/main/kotlin/mimsoft/io/features/checkout/CheckoutResponseDto.kt
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.annotation.JsonInclude
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 data class CheckoutResponseDto(
-    val productCount: Long? = null,
+    val productCount: Int? = null,
     val discountProduct: Long? = null,
     val discountDelivery: Long? = null,
     val promoCode: String? = null,
-    val deliveryPrice: Double? = null,
-    val total: Double? = null
+    val deliveryPrice: Long? = null,
+    val total: Long? = null
 )
